refactor: petition 패키지 리팩토링 (#97)
