+++ src/main/java/com/smunity/api/domain/account/domain/User.java
@@ -2,7 +2,6 @@ package com.smunity.api.domain.account.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonProperty.Access;
-import com.smunity.api.domain.petition.domain.Petition;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
