+++ src/main/java/com/smunity/api/domain/petition/controller/AgreementController.java
@@ -2,6 +2,7 @@ package com.smunity.api.domain.petition.controller;
 
 import com.smunity.api.domain.petition.dto.AgreementDto;
 import com.smunity.api.domain.petition.service.AgreementService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -9,13 +10,10 @@ import java.util.List;
 
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/petitions/{petitionId}/agreements")
 public class AgreementController {
     private final AgreementService agreementService;
-
-    public AgreementController(AgreementService agreementService) {
-        this.agreementService = agreementService;
-    }
 
     @GetMapping()
     public ResponseEntity<List<AgreementDto>> getAgreementUsersByPetitionId(@PathVariable Long petitionId) {
