+++ src/main/java/com/smunity/api/domain/petition/controller/RespondController.java
@@ -2,19 +2,17 @@ package com.smunity.api.domain.petition.controller;
 
 import com.smunity.api.domain.petition.dto.RespondDto;
 import com.smunity.api.domain.petition.service.RespondService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/petitions/{petitionId}/responds")
 public class RespondController {
     private final RespondService respondService;
-
-    public RespondController(RespondService respondService) {
-        this.respondService = respondService;
-    }
 
     @PostMapping()
     public ResponseEntity<RespondDto> createRespond(@PathVariable Long petitionId, @RequestBody RespondDto respondDto, @RequestHeader(value = "X-AUTH-TOKEN") String token) {
