+++ src/main/java/com/smunity/api/domain/petition/dto/AgreementDto.java
@@ -1,9 +1,8 @@
 package com.smunity.api.domain.petition.dto;
 
 import com.smunity.api.domain.account.domain.User;
-import com.smunity.api.domain.petition.domain.Agreement;
-import com.smunity.api.domain.petition.domain.Comment;
-import com.smunity.api.domain.petition.domain.Petition;
+import com.smunity.api.domain.petition.entity.Agreement;
+import com.smunity.api.domain.petition.entity.Petition;
 import lombok.Builder;
 import lombok.Data;
