+++ src/main/java/com/smunity/api/domain/petition/dto/PetitionDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import com.smunity.api.domain.account.domain.User;
-import com.smunity.api.domain.petition.domain.Petition;
+import com.smunity.api.domain.petition.entity.Petition;
 import lombok.Builder;
 import lombok.Data;
