+++ src/main/java/com/smunity/api/domain/petition/dto/RespondDto.java
@@ -1,8 +1,8 @@
 package com.smunity.api.domain.petition.dto;
 
 import com.smunity.api.domain.account.domain.User;
-import com.smunity.api.domain.petition.domain.Respond;
-import com.smunity.api.domain.petition.domain.Petition;
+import com.smunity.api.domain.petition.entity.Respond;
+import com.smunity.api.domain.petition.entity.Petition;
 import lombok.Builder;
 import lombok.Data;
 import java.time.LocalDateTime;
