+++ src/main/java/com/smunity/api/domain/petition/entity/Petition.java
@@ -1,9 +1,10 @@
-package com.smunity.api.domain.petition.domain;
+package com.smunity.api.domain.petition.entity;
 
 import com.smunity.api.domain.account.domain.User;
 import com.smunity.api.global.common.BaseEntity;
 import lombok.*;
 import javax.persistence.*;
+import java.time.LocalDateTime;
 
 
 @Entity
@@ -11,22 +12,32 @@ import javax.persistence.*;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "petitions_comment")
-public class Comment extends BaseEntity {
+@Table(name = "petitions_petition")
+public class Petition extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
+    private String subject;
+
+    @Column(nullable = false)
     private String content;
+
+    @Column(nullable = false)
+    private Integer category;
+
+    @Column(nullable = false)
+    private Boolean anonymous;
+
+    @Column
+    private LocalDateTime end_date;
+
+    @Column(nullable = false)
+    private Integer status;
 
     @ManyToOne
     @JoinColumn(name = "author_id")
     @ToString.Exclude
     private User author;
-
-    @ManyToOne
-    @JoinColumn(name = "petition_id")
-    @ToString.Exclude
-    private Petition petition;
 }
