+++ src/main/java/com/smunity/api/domain/petition/entity/Respond.java
@@ -1,4 +1,4 @@
-package com.smunity.api.domain.petition.domain;
+package com.smunity.api.domain.petition.entity;
 
 import com.smunity.api.domain.account.domain.User;
 import com.smunity.api.global.common.BaseEntity;
@@ -11,18 +11,21 @@ import javax.persistence.*;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "petitions_petition_voter")
-public class Agreement extends BaseEntity {
+@Table(name = "petitions_answer")
+public class Respond extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    @ToString.Exclude
-    private User user;
+    @Column(nullable = false)
+    private String content;
 
     @ManyToOne
+    @JoinColumn(name = "author_id")
+    @ToString.Exclude
+    private User author;
+
+    @OneToOne
     @JoinColumn(name = "petition_id")
     @ToString.Exclude
     private Petition petition;
