+++ src/main/java/com/smunity/api/domain/petition/repository/AgreementRepository.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.petition.repository;
 
-import com.smunity.api.domain.petition.domain.Agreement;
+import com.smunity.api.domain.petition.entity.Agreement;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
