+++ src/main/java/com/smunity/api/domain/petition/repository/RespondRepository.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.petition.repository;
 
-import com.smunity.api.domain.petition.domain.Respond;
+import com.smunity.api.domain.petition.entity.Respond;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
