+++ src/test/java/com/smunity/api/domain/petition/repository/AgreementRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.petition.repository;
 
-import com.smunity.api.domain.petition.domain.Agreement;
+import com.smunity.api.domain.petition.entity.Agreement;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
