+++ src/test/java/com/smunity/api/domain/petition/repository/PetitionRepositoryTest.java
@@ -1,7 +1,6 @@
 package com.smunity.api.domain.petition.repository;
 
-import com.smunity.api.domain.petition.domain.Petition;
-import com.smunity.api.domain.petition.repository.PetitionRepository;
+import com.smunity.api.domain.petition.entity.Petition;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
